feat: :art: Creando Gates sencillas
Creando gates de update y delete

+++ Agenda/app/Http/Controllers/ContactoController.php
@@ -1,42 +1,26 @@
 <?php
 
 namespace App\Http\Controllers;
-
 use App\Models\Contacto;
+use App\Models\User;
+use App\Policies\ContactosPolicy;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class ContactoController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index(Request $request)
-    {
+    public function index(Request $request) {
+        // $contactos = Contacto::where('user_id', Auth::id())->latest()->paginate(5);
         $contactos = Contacto::all();
         return view('contactos.index', compact('contactos'));
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        // $this->authorize('contactos.create');
+    public function create() {
         return view('contactos.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
+    public function store(Request $request) {
         // Validacion de todo rellenado
         $request->validate([
             'name'=> 'required|max:20',
@@ -51,72 +35,37 @@ class ContactoController extends Controller
 
         $contacto->save();
         return redirect()->route('contactos.show', $contacto);
-    }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Models\Contacto  $contacto
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Contacto $contacto)
-    {
+    }
+    public function show(Contacto $contacto) {
         return view('contactos.show',compact('contacto'));
     }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  \App\Models\Contacto  $contacto
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Contacto $contacto)
-    {
-        // $this->authorize('contactos.update',$contacto);
+    public function edit(Contacto $contacto) {
         return view('contactos.edit', compact('contacto'));
     }
+    public function update(Request $request, Contacto $contacto) {
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Contacto  $contacto
-     * @return \Illuminate\Http\Response
-     */
-    public function update(Request $request, Contacto $contacto)
-    {
-        // $this->authorize('update-contacto', $contacto);
-        // if(Gate::denies('update-contacto', $contacto)) {
-        //     abort(403, "No puedes actualizar el contacto");
-        // }
-        // $this->authorize('contactos.update',$contacto);
+        if(Gate::denies('update-contacto', $contacto)) {
+            abort(403);
+        }
 
         // Validacion de todo rellenado
         $request->validate([
             'name'=> 'required|max:20',
-            'numero'=> 'required|max:12',
+            'phone'=> 'required|max:12',
         ]);
 
         $contacto->name = $request->name;
-        $contacto->numero = $request->numero;
+        $contacto->phone = $request->phone;
 
         $contacto->save();
         return redirect()->route('contactos.show', $contacto);
     }
+    public function destroy(Contacto $contacto) {
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  \App\Models\Contacto  $contacto
-     * @return \Illuminate\Http\Response
-     */
-    public function destroy(Contacto $contacto)
-    {
-        // $this->authorize('contactos.delete', $contacto);
-        // if(Gate::denies('destroy-contacto', $contacto)) {
-        //     abort(403, "No puedes eliminar el contacto");
-        // }
-
+        if(Gate::denies('delete-contacto', $contacto)) {
+            abort(403);
+        }
         $contacto->delete();
         return redirect()->route('contactos.index');
     }
